+++ packages/contracts/contracts/core/component/DaoAuthorizableUpgradeable.sol
@@ -9,12 +9,12 @@ import {_auth} from "../../utils/auth.sol";
 
 import {IDAO} from "./../IDAO.sol";
 
-/// @title DaoAuthorizableUpgradable
+/// @title DaoAuthorizableUpgradeable
 /// @author Aragon Association - 2022
 /// @notice An abstract contract providing a meta transaction compatible modifier to authorize function calls through an associated DAO.
 /// This contract provides an `auth` modifier that can be applied to functions in inheriting contracts. The permission to call these functions is managed by the associated DAO.
 /// @dev Make sure to call `__DaoAuthorizable_init` during initialization of the inheriting contract.
-///      This contract is compatible with meta transactions through OZ's `ContextUpgradable`.
+///      This contract is compatible with meta transactions through OZ's `ContextUpgradeable`.
 abstract contract DaoAuthorizableUpgradeable is Initializable, ContextUpgradeable {
     /// @notice The associated DAO managing the permissions of inheriting contracts.
     IDAO internal dao;